+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Print BLE download progress after chunk is complete instead of before.
+
 ## [1.0.0-alpha.2] - 2021-04-08
 ### Added
 - Check Pybricks protocol version when connecting to Bluetooth Low Energy devices.